+++ changelog.md
@@ -1,6 +1,6 @@
 # 1.5.2
 
-- Temporarily remove the alert on quests due to a crash
+- Probably fix a crash
 
 # 1.5.1